requirements.txt
aiofiles
aiohttp
aiosignal
altair
annotated-types
anyio
asttokens
attrs
certifi
charset-normalizer
click
colorama
comm
contourpy
cycler
dataclasses-json
debugpy
decorator
distro
executing
fastapi
ffmpy
filelock
fonttools
frozenlist
fsspec
gradio
gradio_client
greenlet
h11
httpcore
httpx
huggingface-hub
idna
importlib_resources
ipykernel
ipython
jedi
Jinja2
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
kiwisolver
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-neo4j
langchain-text-splitters
langsmith
markdown-it-py
MarkupSafe
marshmallow
matplotlib
matplotlib-inline
mdurl
multidict
mypy-extensions
neo4j
nest-asyncio
numpy
openai
orjson
packaging
pandas
parso
pillow
platformdirs
prompt-toolkit
psutil
pure-eval
pydantic
pydantic_core
pydub
Pygments
pyparsing
pyprojroot
python-dateutil
python-dotenv
python-multipart
pytz
PyYAML
pyzmq
referencing
regex
requests
rich
rpds-py
ruff
semantic-version
shellingham
six
sniffio
spacy
SQLAlchemy
stack-data
starlette
tenacity
tiktoken
tomlkit
toolz
tornado
tqdm
traitlets
typer
typing-inspect
typing_extensions
tzdata
urllib3
uvicorn
wcwidth
websockets
yarl
